Validate amount against Liquid Network payment limits (#64)
* Address typos on PaymentError's

* Disable description fields until it's passable to payment data

* Rename liquidMinimumPaymentAmountSat & move maxPaymentAmount to constants

* Convert PaymentError amount types to int

* Validate payment against payment limits & balance

* Fetch lightning limits on initializing LnUrlCubit

* Validate LnUrl payments against ln payment limits

* Wait for lightning limits to be fetched on payment pages
* Remove minPaymentAmountSat from constants
* Comment-out unused _descriptionController
* Differentiate error messages for LNURL request data limits & LN payment limits

+++ lib/cubit/chain_swap/chain_swap_cubit.dart
@@ -115,10 +115,10 @@ class ChainSwapCubit extends Cubit<ChainSwapState> {
   ) {
     var limits = outgoing ? onchainLimits.send : onchainLimits.receive;
     if (amount > limits.maxSat) {
-      throw PaymentExceededLimitError(limits.maxSat);
+      throw PaymentExceededLimitError(limits.maxSat.toInt());
     }
     if (amount < limits.minSat) {
-      throw PaymentBelowLimitError(limits.minSat);
+      throw PaymentBelowLimitError(limits.minSat.toInt());
     }
   }
 }
